requirements.txt
django<1.12
djangorestframework<3.8
psycopg2-binary<2.8
python-dateutil<2.8
requests<2.2
gunicorn==19.6.0
whitenoise==3.2
django-rest-swagger==2.2.0
pika=0.12.0
